+++ services/details/internal/adapter/outbound/messaging/noop.go
@@ -1,4 +1,4 @@
-package kafka
+package messaging
 
 import (
 	"context"
